Fixing styling + Mobile main image
