+++ assets/css/main.scss
@@ -803,9 +803,10 @@ display: none;
     background-position: center;
     background-size: cover;
     // background-color: #FAFAFA;
-    background-image: url("/images/cypherKeys.gif");
+    background-image: url("/images/mobile.gif");
     @include breakpoint(phone) {
       height: 90vh;
+      background-image: url("/images/cypherKeys.gif");
     }
     &:before {
       content: '';
