+++ package.json
@@ -43,6 +43,7 @@
     "react-rater": "^5.1.1",
     "react-redux": "^7.2.1",
     "react-select": "^3.1.0",
+    "react-toastify": "^9.1.3",
     "react-transition-group": "^4.4.5",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
